requirements.txt
absl-py==1.1.0
antlr4-python3-runtime==4.8
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bitarray==2.6.0
bleach==5.0.1
blis==0.4.1
breadability==0.1.20
catalogue==2.0.7
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==7.1.2
colorama==0.4.5
cycler==0.11.0
cymem==2.0.6
Cython==0.29.32
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
fairseq~=0.12.2
fastapi~=0.81.0
fastBPE==0.1.0
fastjsonschema==2.15.3
filelock==3.7.1
fonttools==4.33.3
gdown==4.5.3
graphviz==0.20
h11==0.13.0
huggingface-hub==0.8.1
hydra-core==1.0.7
icecream==2.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipykernel==6.15.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.6.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
langcodes==3.3.0
lexrank~=0.1.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-service==2.4.0
murmurhash==1.0.7
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.6.3
nltk~=3.7
notebook==6.4.12
omegaconf==2.0.6
packaging==21.3
pandas~=1.3.5
pandocfilters==1.5.0
parso==0.8.3
path==16.4.0
path.py==12.5.0
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plac==1.1.3
platformdirs==2.5.2
portalocker==2.5.1
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pycountry==22.3.5
pycparser==2.21
pydantic~=1.7.4
Pygments==2.12.0
pyngrok~=5.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
pytextrank==3.2.3
python-crfsuite==0.9.8
python-dateutil==2.8.2
pytz==2022.1
pyvi==0.1.1
PyYAML==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.6.2
requests~=2.28.1
rouge~=1.0.1
rouge-metric==1.0.1
rouge-score==0.0.4
sacrebleu==2.2.0
scikit-learn~=1.0.2
scipy~=1.7.3
Send2Trash==1.8.0
sklearn-crfsuite==0.3.6
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.0.8
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
starlette==0.19.1
summa==1.2.0
sumy~=0.10.0
tabulate==0.8.10
terminado==0.15.0
thinc==8.0.17
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
torch~=1.13.0
torchaudio==0.8.2
torchmetrics==0.9.2
torchvision==0.9.2
tornado==6.2
tqdm~=4.64.0
traitlets==5.3.0
transformers~=4.21.0
typer==0.3.2
typing-extensions==3.10.0.2
underthesea~=1.3.4
underthesea_core==0.0.4a10
Unidecode==1.3.4
uritools==4.0.0
urlextract==1.6.0
urllib3==1.26.9
uvicorn==0.18.3
vncorenlp~=1.0.3
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
zipp==3.8.0

numpy~=1.22.4
datasets~=2.6.1
setuptools~=60.2.0